refactor(provider): migrate Gemini to provider-centric adapter architecture
- unify GeminiProvider as a single logical provider identity
- introduce WebAPI and Playwright backend adapters
- remove Playwright as a top-level provider from ProviderFactory
- make ProviderFactory resolve provider identities only
- centralize Gemini adapter selection within GeminiProvider
- extract shared Gemini logic into reusable utilities
- preserve backward compatibility for playwright/ model aliases
- add fail-fast Gemini backend configuration validation
- stabilize /v1/models output across backend implementations
- eliminate provider package circular import hazards
- update architecture and API documentation for Provider/Adapter model
