+++ src/app/services/providers/gemini/__init__.py
@@ -1,0 +1,5 @@
+# src/app/services/providers/gemini/__init__.py
+
+# Note: This package init is intentionally thin to avoid circular dependencies
+# during initialization of the provider architecture.
+# Shared symbols should be imported from their respective modules.
